Reimplement gossip round

+++ pkg/internal/httpmessage/messages.go
@@ -6,7 +6,7 @@ import "github.com/rstefan1/bimodal-multicast/pkg/internal/buffer"
 type HTTPGossip struct {
 	Addr        string              `json:"http_gossip_addr"`
 	Port        string              `json:"http_gossip_port"`
-	RoundNumber int                 `json:"http_gossip_round_number"`
+	RoundNumber int64               `json:"http_gossip_round_number"`
 	Digests     buffer.DigestBuffer `json:"http_gossip_digests"`
 }
 
@@ -14,7 +14,7 @@ type HTTPGossip struct {
 type HTTPSolicitation struct {
 	Addr        string              `json:"http_solicitation_addr"`
 	Port        string              `json:"http_solicitation_port"`
-	RoundNumber int                 `json:"http_solicitation_round_number"`
+	RoundNumber int64               `json:"http_solicitation_round_number"`
 	Digests     buffer.DigestBuffer `json:"http_solicitation_digests"`
 }
